Fix pointer to function comparision in KASSERT.
Pointer to function is not at 100% compatible with void*.
Void* is a type for data, but pointer to function is address to text section.
However we can safety compare those 2 things with simple cast.

This fix is adding full pointer detection and add auto cast to void*.

Issue has been reported by Kamil Kiełbasa